Smallest multiple in a sequential range of numbers
Used a brute force approach due to the size of range needed. Increasing range size has an o(n^2) increase in time complexity for algorithm. This has less to do with code implementation than a rule of numbers. With a range of 1 - 10, I can find the number with a loop counter limit of 10,000. 1 - 20 requires me to increase the limit x5 magnitude to 100,000,000,00.