Find prime at position x
Creating primes using sieve method. Pretty simple problem since we can just go to position x in our primes array to grab our prime. I like to visualize the list of primes, but we don't need to actually keep track of all of them. We can use a counter and our currentPrime instead. On the flip side, if we are calculating primes via positions multiple times, memoizing the primes array would be the best way to go.