Power Digit Sums
This was a fun problem utilizing map and reduce. Reduce is becoming a very used function in these problems! The nuance to this problem is that number doesn't hold the proper amount of precision to calculate the sum of digits of a very large number. There is a BigInt type available to js. So our steps are to conver to bigInt, convert to string, convert to array, map array to numbers and reduce to find sum of digits.