Pythagorean Triplet Product
Quick and simple problem - we are trying different combinations of a and b to solve for c. Once we have c, we check a < b < c and a^2 + b^2 = c^2. Since we have a nested for loop, runs in exponential time. It's interesting to generate triplets for different x sum values (a + b + c = x). Lots of x values don't produce valid pythagorean triplets.