Large Sum
Calculated the first 10 digits of a large sum of 100 50 digit numbers. This problem involved traversing a grid top to bottom and right to left. For each cell we visit in a column, add value to running sum. When we go to a new column, grab remainder and add to sum first.