Lattice Paths
Found the largest number of paths given an NxN grid. Each point not along the left or top edge is calculated by the values of the points to the left and above it. This way we only need to initialize a 2d array and then we can visit every point once to fill out its value. Return value at bottom right corner of grid.