Maximum Path Sum I
Found the maximal total from top to bottom of a triangle with 20 rows, starting at 1 row 1 and increasing by1 for each row. Great application of trees. Started by taking each number on the bottom row as the root of a new tree. Each left/right path calculated set the new value as the running sum of the paths so far. Doing this, each addition to a left or right node does a check against a maxSum. Each left or right child was added to a queue where its children could be calculated. Fun way around recursion.  When the loops done, we will have visited every path from the bottom up.