Maximal Path Sum II
Find the maximal path sum from top to bottom of a triangle with 100 rows on the bottom. Great problem - I tried to solve this using divide and conquer and dynamic programming as I noticed repeating paths down trees that we could hold the sum for. Trying to solve this with trees works for smaller numbers  and is an effective greedy strategy as it visits every path possible, but not with anything bigger than 40 rows of numbers at the bottom. Can be solved by inductive reasoning, solve the base case first. Start at row 1 to end of triangle, find maximum possible values per cell visited and rewrite the node value as its greatest possible value so far. This way, every new row we visit, we can rewrite each node as its greatest value. Then we just need to check against a max sum every time we rewrite a node.