Amicable Numbers
Find the sum of the amicable numbers below 10000. Define d(n) as the sum of proper divisors. If d(a) = b and d(b) =a and a !== b we have amicable numbers a and b. Calculating divisors for small numbers is trivial, but runs in o(n^2) runtime. Finding amicable numbers is faster - had to make sure to hash amicable numbers so we dont double count.