Reciprocal Cycles
Fun one. JS has about 20 digits of decimal precision which is significantly less than the required precision we need. Say hello to Long Division! We know that every decimal is either repeating or non repeating. If at any point during long division, we get a remainder of 0, we are done and there are no repeating digits. Otherwise, for each remainder, we hash it along with the loop counter we are on. This way, when we calculate our remainders, if we find one in our hash, we know we have found the beginning of our repeating sequence. Return the proper portion of decimal digits that belong to the repeating subset. Doesn't run in great time, o(n^2). Too optimize, I would first look at the decimal digits created. Are there repeating blocks of numbers that we could cache outside of the d calculation? Besides that, we could definitely not calculate certain d . For instance, we know d that are multiples of 10 will always have low repeat counts.