Number Spiral Diagonals
Fun problem - first passthrough I thought we could use a grid and figure out how to fill it up from the middle in a spiral pattern. I could use the direction to calculate which way to run the new row. Don't need to do that, much easier just keeping track of the number you're on, direction, row size, and current sum. Using this, we can use nested loops to calculate how much more we need to move to find the next diagonal. Every full path in a direction ends with some type of diagonal we know. Going right is a special case as well as going right on the last spiral.