Circular Primes
Remembered how to sieve - was pretty fun. Simple problem - generate primes up to limit + 1. For each prime, we rotate it n - 1 times where n is amount of digits in prime. Rotation is easy if we convert into string, then array, shift top off, push to back, and check if that num is in primes hash. Return sum of these circular primes