Truncatable Primes
Another prime problem - so we use sieve to generate primes. Then we simply loop through primes and do left and right rotations. At each rotation, we check against our hash of primes. If at anytime we don't find a rotation in the hash, we return false for that number. Sum the primes that return true. Runs fast, almost linear time. Sieving is o(nlogn) which is the longest operation